Ai Translator App/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Werkzeug
requests
gTTS
pytesseract
Pillow
langdetect